* Cleaned BenchmarkCoreObject target to inherit all its build settings
* Turned on few Clang-specific warnings about implicit conversions in all targets

* Added -Wextra to all targets except TestCoreObject and BenchmarkCoreObject

* Turned on ARC by default (BasicPersistence still uses manual memory management)


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@9913 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Benchmark/TestSQLiteStorePerformance.m
@@ -179,7 +179,7 @@ static int itemChangedAtCommit(int i)
 {
     NSMutableDictionary *dict = [NSMutableDictionary dictionaryWithCapacity: numChildren+1];
     
-    for (int i=0; i<numChildren; i++)
+    for (NSUInteger i=0; i<numChildren; i++)
     {
         @autoreleasepool {
             COMutableItem *item = [COMutableItem item];
